+++ example/download.dart
@@ -5,7 +5,8 @@ import 'package:dio/dio.dart';
 main() async {
   var dio = new Dio();
 
-  dio.onHttpClientCreate=(HttpClient client){
+  (dio.httpClientAdapter as DefaultHttpClientAdapter).onHttpClientCreate =
+      (HttpClient client) {
     client.idleTimeout = new Duration(seconds: 0);
   };
 
@@ -13,19 +14,25 @@ main() async {
   // var url = "http://download.dcloud.net.cn/HBuilder.9.0.2.macosx_64.dmg";
 
   // This is a image, about 4KB
-  //var url = "https://flutter.io/assets/flutter-lockup-4cb0ee072ab312e59784d9fbf4fb7ad42688a7fdaea1270ccf6bbf4f34b7e03f.svg";
-  var url="https://github.com/wendux/tt";
+  var url = "https://flutter.io/assets/flutter-lockup-4cb0ee072ab312e59784d9fbf4fb7ad42688a7fdaea1270ccf6bbf4f34b7e03f.svg";
+  //var url = "https://github.com/wendux/tt"; //404
   try {
-    Response response=await dio.download(url,
+    Response response = await dio.download(
+      url,
       "./example/flutter.svg",
       onProgress: (received, total) {
-        print((received / total * 100).toStringAsFixed(0) + "%");
+        if (total != -1) {
+          print((received / total * 100).toStringAsFixed(0) + "%");
+        }
       },
-      options: Options(receiveDataWhenStatusError: false)
+      cancelToken: CancelToken(),
+      options: Options(
+        //receiveDataWhenStatusError: false,
+        headers: {HttpHeaders.acceptEncodingHeader: "*"},
+      ),
     );
-    print(response.statusCode);
-  } on DioError catch (e) {
+    print("download succeed!");
+  } catch (e) {
     print(e.response.data);
   }
-  print("download succeed!");
 }