+++ example/proxy.dart
@@ -5,7 +5,7 @@ main() async {
   var dio = new Dio();
   //dio.options.connectTimeout = 2000;
   // More about HttpClient proxy topic please refer to Dart SDK doc.
-  dio.onHttpClientCreate = (HttpClient client) {
+  (dio.httpClientAdapter as DefaultHttpClientAdapter).onHttpClientCreate = (HttpClient client) {
     client.findProxy = (uri) {
       //proxy all request to localhost:8888
       return "PROXY localhost:8888";