+++ lib/dio.dart
@@ -16,3 +16,4 @@ export 'src/cancel_token.dart';
 export 'src/upload_file_info.dart';
 export 'src/interceptors/log.dart';
 export 'src/interceptors/cookie_mgr.dart';
+export 'src/adapter.dart';